rename back index.js to eventsource.js
