+++ lib/eventsource/eventsource.js
